Add OptionalNumberTranslationFilter
Move OptionalNumberTranslationFilter from CensusDataFilter to specific filter.
